+++ plugin/web/extensions/sidebar/modals/datacatalog/api/api.ts
@@ -14,6 +14,8 @@ export type { DataCatalogItem, DataCatalogGroup, DataCatalogTreeItem };
 
 export type RawDataCatalogTreeItem = RawDataCatalogGroup | RawDataCatalogItem;
 
+export type { GroupBy } from "./path";
+
 export type RawDataCatalogGroup = {
   id: string;
   name: string;
