+++ server/datacatalog/datacatalogutil/common.go
@@ -4,6 +4,8 @@ import (
 	"fmt"
 	"net/url"
 	"path"
+	"regexp"
+	"strconv"
 	"strings"
 
 	"github.com/eukarya-inc/jpareacode"
@@ -84,4 +86,42 @@ func IsLayerSupported(format string) bool {
 		return true
 	}
 	return false
+}
+
+var reReiwa = regexp.MustCompile(`令和([0-9]+?)年度`)
+
+func YearInt(y string) (year int) {
+	if ym := reReiwa.FindStringSubmatch(y); len(ym) > 1 {
+		yy, _ := strconv.Atoi(ym[1])
+		if yy > 0 {
+			year = yy + 2018
+		}
+	} else if yy, err := strconv.Atoi(strings.TrimSuffix(strings.TrimSuffix(y, "度"), "年")); err == nil {
+		year = yy
+	}
+	return year
+}
+
+func FormatTypeEn(f string) string {
+	if f == "3D Tiles" {
+		return "3dtiles"
+	}
+	return strings.ToLower(f)
+}
+
+const Zenkyu = "全球データ"
+
+func NormalizePref(pref string) (string, int) {
+	if pref == "全球" || pref == "全国" {
+		pref = Zenkyu
+	}
+
+	var prefCode int
+	if pref == Zenkyu {
+		prefCode = 0
+	} else {
+		prefCode = jpareacode.PrefectureCodeInt(pref)
+	}
+
+	return pref, prefCode
 }
