+++ server/datacatalog/item_plateu.go
@@ -2,7 +2,7 @@ package datacatalog
 
 import (
 	"github.com/eukarya-inc/reearth-plateauview/server/datacatalog/plateauv2"
-	"github.com/reearth/reearthx/util"
+	"github.com/samber/lo"
 )
 
 type PlateauItem plateauv2.CMSItem
@@ -18,9 +18,9 @@ func (i PlateauItem) DataCatalogs() []DataCatalogItem {
 	if c.Year == 0 {
 		return nil
 	}
-	return util.Map(plateauv2.CMSItem(i).AllDataCatalogItems(c), dataCatalogItemFromPlateauV2)
+	return lo.Map(plateauv2.CMSItem(i).AllDataCatalogItems(c), dataCatalogItemFromPlateauV2)
 }
 
-func dataCatalogItemFromPlateauV2(i plateauv2.DataCatalogItem) DataCatalogItem {
+func dataCatalogItemFromPlateauV2(i plateauv2.DataCatalogItem, _ int) DataCatalogItem {
 	return DataCatalogItem(i)
 }
