+++ server/datacatalog/plateauv2/dataset_test.go
@@ -1,4 +1,4 @@
-package datacatalog
+package plateauv2
 
 import (
 	"testing"
